team_collaboration_api/requirements.txt
# Team Collaboration API
Flask>=3.0.0,<4.0.0
Flask-SQLAlchemy>=3.1.0,<4.0.0
Flask-JWT-Extended>=4.6.0,<5.0.0
Flask-Cors>=4.0.0,<5.0.0
Flask-SocketIO>=5.3.0,<6.0.0
Flask-Caching>=2.1.0,<3.0.0
redis>=5.0.0,<6.0.0
celery[redis]>=5.3.0,<6.0.0
marshmallow>=3.21.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
flasgger>=0.9.7.1,<1.0.0
eventlet>=0.35.0,<1.0.0
email-validator>=2.0.0,<3.0.0

# Testing
pytest>=8.0.0,<9.0.0
pytest-cov>=5.0.0,<7.0.0
pytest-flask>=1.3.0,<2.0.0
